chore(workspace): apply cargo clippy --fix (auto-resolvable lints)
Runs `cargo clippy --fix --workspace --all-targets --allow-dirty` to
auto-resolve the trivial subset of the pre-pivot clippy debt:

- needless borrows on format!() args
- push_str of single-character literals
- `or_insert_with(HashSet::new)` → `or_default()`
- `.unwrap_or_else(Default::default)` → `.unwrap_or_default()`
- useless `vec![]` for short fixed arrays
- needless lifetime annotations
- format!() arg inlining where applicable

Mostly in `crates/rts-core/`; a handful of touch-ups in `rts-mcp` and
`rts-bench` tests. Behaviour-preserving by construction (rustc/clippy
guarantee for `--fix`). `cargo test --workspace` → 360 passed, unchanged.

Remaining clippy debt (~75 lints needing manual judgment) is deferred
and covered by the matching CI relaxation in the next commit.

+++ crates/rts-core/src/advanced_parallel.rs
@@ -253,8 +253,8 @@ impl Worker {
                 let _total_tasks_u32 = total_tasks as u32;
                 let duration_u32 = duration.as_millis() as u32;
                 stats_guard.average_task_duration = Duration::from_millis(
-                    ((current_avg.as_millis() as u64 * (total_tasks - 1) + duration_u32 as u64)
-                        / total_tasks) as u64,
+                    (current_avg.as_millis() as u64 * (total_tasks - 1) + duration_u32 as u64)
+                        / total_tasks,
                 );
             }
         }
